+++ requests/booking.go
@@ -4,16 +4,16 @@ type BookingReq struct {
 	Mbmid         string    `json:"mbmid"`
 	BookingNumber string    `json:"booking_number"`
 	BookingDate   string    `json:"booking_date"`
-	PayAmt        int       `json:"pay_amount"`
+	PayAmt        float32   `json:"pay_amount"`
 	Emoney        int       `json:"emoney"`
 	PayMethod     string    `json:"payment_method"`
 	Trf           []bookTrf `json:"trf"`
 }
 
 type bookTrf struct {
-	TrfID     int    `json:"trf_id"`
-	TrfQty    int    `json:"trf_qty"`
-	TrfType   string `json:"trf_trftype"`
-	TrfAmount int    `json:"trf_amount"`
-	TrfTotal  int    `json:"trf_total"`
+	TrfID     int     `json:"trf_id"`
+	TrfQty    int     `json:"trf_qty"`
+	TrfType   string  `json:"trf_trftype"`
+	TrfAmount float32 `json:"trf_amount"`
+	TrfTotal  float32 `json:"trf_total"`
 }
